drm/i915/gvt: force to active the high-performance mode during vGPU busy
With the RPS interrupt, KMD can adjust the GPU frequency dynamically for
power saving. It works well in the non-virtualized environment, but there
is more latency imported by VMM and virtual interrupt handler which may
break the RPS policy work model, and GPU works in inefficient mode. Here
we force to active the high-performance mode when detect vgpu is busy until
the GPU runs into idle.

Change-Id: I7bc506b811a94314f068a7891b5e250a6f3f7162
Tracked-On: projectacrn/acrn-hypervisor/issues/2227
Signed-off-by: Weinan Li <weinan.z.li@intel.com>
Signed-off-by: Min He <min.he@intel.com>
Reviewed-by: Zhao Yakui <yakui.zhao@intel.com>
Tracked-On: PKT-1642