vhm: add ioeventfd support for ACRN hypervisor service module
ioeventfd which is based on eventfd, intends to glue vhm module and
other modules who are interested in guest IOs.

Each ioeventfd registered by userspace can map a PIO/MMIO range of
the guest to eventfd, and response to signal the eventfd when get the
in-range IO write from guest. Then the other side of eventfd can be
notified to process the IO request.

Now we only use the ioeventfd to listen virtqueue's kick register,
there are some limitations:
  1) Length support can only be 1, 2, 4 or 8
  2) Only support write operation, read will get 0
  3) Same address, shorter length writing can be handled with the
     integral data matching

Tracked-On: projectacrn/acrn-hypervisor#1329
Signed-off-by: Shuo Liu <shuo.a.liu@intel.com>
Reviewed-by: Yu Wang <yu1.wang@intel.com>
Reviewed-by: Zhao Yakui <yakui.zhao@intel.com>