drm/i915/gvt: add PIPEDSL to pvmmio trap list to read HW PIPEDSL register for guest
The PipeDSL register will be read in irq handler. But currently it
is accessed based on write-protected mode after PVMMIO is enabled.
This patch added PIPEDSL registers to the pvmmio trap list so that guest
will be able to read the HW PIPEDSL registers. This patch can help to
reduce the time consumption of vblank irq handler in guest.

Change-Id: Ie9d2fa0125db9dadd5475fa8aeb036cde2adb90d
Tracked-On: projectacrn/acrn-hypervisor#2597
Signed-off-by: He, Min <min.he@intel.com>
Reviewed-by: Zhao Yakui <yakui.zhao@intel.com>
Tracked-On: PKT-1750